requirements.txt
boto==2.38.0
Flask==0.10.1
itsdangerous==0.24
Jinja2==2.7.3
MarkupSafe==0.23
Pillow==2.8.1
Werkzeug==0.10.4
